major rework on backend logic - seems to work locally
