+++ setup.py
@@ -1,5 +1,5 @@
 import pathlib
-from setuptools import setup, find_packages
+from setuptools import setup, find_packages, find_namespace_packages
 
 HERE = pathlib.Path(__file__).parent
 README = (HERE / "README.md").read_text()
@@ -14,13 +14,15 @@ setup(
     author='EIDOSlab',
     author_email='eidoslab@di.unito.it',
     license='MIT',
-    package_dir={
-                'torchstain': 'torchstain',
-                'torchstain.torch': 'torchstain/torch',
-                'torchstain.tf': 'torchstain/tf'},
-    #packages=['torchstain', 'torchstain.torch',
-    #          'torchstain.tf'],
+    #package_dir={
+    #            'torchstain': 'torchstain',
+    #            'torchstain.torch': 'torchstain/torch',
+    #            'torchstain.tf': 'torchstain/tf'},
+    #packages=['torchstain', 'torchstain.torch',],
+    #          #'torchstain.tf'],
     packages=find_packages(exclude=('tests')),
+    #packages=find_namespace_packages(where='torchstain'),
+    #packages=["torchstain", "torchstain.torch"],
     zip_safe=False,
     install_requires=[
         'numpy'
