+++ tests/test_normalizers.py
@@ -20,13 +20,13 @@ T = transforms.Compose([
 t_to_transform = T(to_transform)
 
 # initialize normalizers for each backend and fit to target image
-normalizer = torchstain.torch.normalizers.MacenkoNormalizer(backend='numpy')
+normalizer = torchstain.base.normalizers.macenko_normalizer.MacenkoNormalizer(backend='numpy')
 normalizer.fit(target)
 
-torch_normalizer = torchstain.torch.normalizers.MacenkoNormalizer(backend='torch')
+torch_normalizer = torchstain.base.normalizers.macenko_normalizer.MacenkoNormalizer(backend='torch')
 torch_normalizer.fit(T(target))
 
-tf_normalizer = torchstain.tf.normalizers.MacenkoNormalizer(backend='tensorflow')
+tf_normalizer = torchstain.base.normalizers.macenko_normalizer.MacenkoNormalizer(backend='tensorflow')
 tf_normalizer.fit(T(target))
 
 # transform
