+++ torchstain/base/__init__.py
@@ -1,0 +1,1 @@
+import torchstain.base.normalizers.macenko_normalizer
