--- torchstain/tf/__init__.py
@@ -1,1 +1,0 @@
-from .normalizers import MacenkoNormalizer
