--- torchstain/tf/utils/__init__.py
@@ -1,3 +1,0 @@
-from torchstain.tf.utils.cov import cov_tf
-from torchstain.tf.utils.percentile import percentile_tf
-from torchstain.tf.normalizers.solveLS import solveLS
