+++ torchstain/tf_backend/__init__.py
@@ -1,0 +1,2 @@
+import torchstain.tf_backend.normalizers.tensorflow_macenko_normalizer
+
