+++ torchstain/tf_backend/normalizers/__init__.py
@@ -1,0 +1,1 @@
+import torchstain.tf_backend.normalizers.tensorflow_macenko_normalizer
