--- torchstain/torch/normalizers/__init__.py
@@ -1,1 +1,0 @@
-from .macenko_normalizer import MacenkoNormalizer
