--- torchstain/torch/normalizers/he_normalizer.py
@@ -1,6 +1,0 @@
-class HENormalizer:
-    def fit(self, target):
-        pass
-
-    def normalize(self, I, **kwargs):
-        raise Exception('Abstract method')
