+++ torchstain/torch_backend/__init__.py
@@ -1,0 +1,1 @@
+import torchstain.torch_backend.normalizers.torch_macenko_normalizer
