+++ torchstain/torch_backend/normalizers/torch_macenko_normalizer.py
@@ -1,6 +1,6 @@
 import torch
-from torchstain.torch.normalizers.he_normalizer import HENormalizer
-from torchstain.torch.utils import cov, percentile
+from torchstain.base.normalizers.he_normalizer import HENormalizer
+from torchstain.torch_backend.utils import cov, percentile
 
 """
 Source code ported from: https://github.com/schaugf/HEnorm_python
