+++ torchstain/torch_backend/utils/__init__.py
@@ -1,0 +1,2 @@
+from torchstain.torch_backend.utils.cov import cov
+from torchstain.torch_backend.utils.percentile import percentile
